refactor:z-index 범주화, 각종 수정 및 폴더구조 변경
