+++ src/App.jsx
@@ -1,6 +1,6 @@
 import { BrowserRouter, Routes, Route, Navigate } from "react-router-dom";
-import Nav from "./components/Nav";
-import Items from "./pages/Items";
+import Nav from "./layouts/Nav";
+import ItemsPage from "./pages/ItemsPage/ItemsPage";
 import AddItemPage from "./pages/AddItemPage/AddItemPage";
 
 const App = () => {
@@ -9,7 +9,7 @@ const App = () => {
       <Nav />
       <Routes>
         <Route path="/" element={<Navigate to="/items" replace />} />
-        <Route path="/items" element={<Items />} />
+        <Route path="/items" element={<ItemsPage />} />
         <Route path="/additem" element={<AddItemPage />} />
       </Routes>
     </BrowserRouter>
