+++ src/layouts/Nav.module.css
@@ -10,7 +10,7 @@
   margin: 0 auto;
   border-bottom: 1px solid var(--color-border-grey);
   gap: var(--space-sm);
-  z-index: 999;
+  z-index: var(--z-index-navigation);
 }
 
 .link-container {
