+++ src/pages/AddItemPage/components/Field.jsx
@@ -1,5 +1,4 @@
-import { useEffect } from "react";
-import useField from "../../../components/hooks/useField";
+import useField from "../../../hooks/useField";
 
 const Field = ({ label, id, type, placeholder, onSaveData }) => {
   const {
