+++ src/pages/AddItemPage/components/FileField.module.css
@@ -5,6 +5,7 @@
 
 .imgContainer {
   display: flex;
+  align-items: center;
   gap: 10px;
   margin: var(--space-md) 0;
 }
@@ -13,7 +14,6 @@
   display: flex;
   flex-direction: column;
   justify-content: center;
-  align-items: center;
   flex: 1 1 50%;
   aspect-ratio: 1/1;
   border-radius: 12px;
